back out int32 in interfaces

+++ godiff/difference_test.go
@@ -31,33 +31,33 @@ func TestWriteInts(t *testing.T) {
 
 func TestDiffSets_new(t *testing.T) {
 	ds := newDiffSet(3, []int32{0, 1})
-	expect(t, "k", ds.k, int32(3))
+	expect(t, "k", ds.k, 3)
 	expect(t, "v", ds.v, int32(7))
 	expect(t, "trials", ds.trials, int64(0))
-	expect(t, "targetDepth", ds.targetDepth, int32(0))
-	expect(t, "current", ds.current, int32(2))
+	expect(t, "targetDepth", ds.targetDepth, 0)
+	expect(t, "current", ds.current, 2)
 	expect(t, "s", ds.s, []int32{0, 1, 0})
 }
 
 func TestDiffSets_Find3(t *testing.T) {
 	ds := newDiffSet(3, []int32{0, 1})
 	ds.Find(nil)
-	expect(t, "k", ds.k, int32(3))
+	expect(t, "k", ds.k, 3)
 	expect(t, "v", ds.v, int32(7))
 	expect(t, "trials", ds.trials, int64(1))
-	expect(t, "targetDepth", ds.targetDepth, int32(0))
-	expect(t, "current", ds.current, int32(3))
+	expect(t, "targetDepth", ds.targetDepth, 0)
+	expect(t, "current", ds.current, 3)
 	expect(t, "s", ds.s, []int32{0, 1, 3})
 }
 
 func TestDiffSets_Find4(t *testing.T) {
 	ds := newDiffSet(4, []int32{0, 1})
 	ds.Find(nil)
-	expect(t, "k", ds.k, int32(4))
+	expect(t, "k", ds.k, 4)
 	expect(t, "v", ds.v, int32(13))
 	expect(t, "trials", ds.trials, int64(4))
-	expect(t, "targetDepth", ds.targetDepth, int32(0))
-	expect(t, "current", ds.current, int32(4))
+	expect(t, "targetDepth", ds.targetDepth, 0)
+	expect(t, "current", ds.current, 4)
 	expect(t, "s", ds.s, []int32{0, 1, 3, 9})
 }
